requirements.txt
# 核心依賴
numpy>=1.19.0
scipy>=1.6.0
pandas>=1.2.0
matplotlib>=3.3.0
seaborn>=0.11.0
psutil>=5.8.0
sympy>=1.8

# 檔案處理
cryptography>=3.4.0
pillow>=8.0.0
opencv-python>=4.5.0
